refactor: update css styles
