+++ src/components/Modal/Modal.css
@@ -1,0 +1,76 @@
+.modal {
+  position: fixed;        /* поверх всего */
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;          /* чтобы центрировать */
+  align-items: center;
+  justify-content: center;
+  z-index: 1000;          /* поверх всего остального */
+}
+
+.modal-content {
+  position: relative;
+  background: #222;
+  padding: 20px;
+  border-radius: 10px;
+  width: 500px;
+  min-height: 500px;
+  max-height: 1000px;
+  box-shadow: 0 4px 12px rgba(0,0,0,0.3);
+  z-index: 1001;
+}
+
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0,0,0,0.5); /* затемнение */
+  z-index: 999;
+}
+
+.close-button {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  background: transparent;
+  border: none;
+  font-size: 20px;
+  cursor: pointer;
+  color: white;
+}
+
+form {
+  padding: 30px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.error {
+  color: #ff5c5c;
+  font-size: 14px;
+  min-height: 28px;
+  margin-top: 4px;
+}
+
+.form-row {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  column-gap: 20px;
+}
+
+.form-field {
+  display: flex;
+  flex-direction: row;
+}
+
+.password-container {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
